% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/momentchi2.R
\docType{package}
\name{momentchi2}
\alias{momentchi2}
\alias{momentchi2-package}
\title{momentchi2: A package containing a collection of moment-matching methods for computing the cdf of a weighted sum of chi-squared random variables}
\description{
The \code{momentchi2} package provides implementations of moment-matching (equivalently, cumulant-matching)
methods for computing the cumulative distribution function (cdf) of a positively-weighted sum of chi-squared random variables.
}
\details{
Methods implemented include the Satterthwaite-Welch (SW), Hall-Buckley-Eagleson (HBE), Wood F (WF) and
Lindsay-Pilla-Basak (LPB4) methods. Note that the chi-squared random variables are assumed to have degree 1.
}
\seealso{
Those interested in weighted sums of chi-squared random variables are encouraged to also have a look at the excellent \code{CompQuadForm} R package, which provides implementations of several iterative methods for computing the cdf of a weighted sum of (potentially noncentral) chi-squared random variables.
}

